+++ lizmap/forms/dataviz_edition.py
@@ -88,6 +88,7 @@ class DatavizEditionDialog(BaseEditionDialog, CLASS):
 
         self.layer.setFilters(QgsMapLayerProxyModel.VectorLayer)
         self.layer.layerChanged.connect(self.current_layer_changed)
+        self.layer.layerChanged.connect(self.check_layer_wfs)
 
         self.x_field.setLayer(self.layer.currentLayer())
 
@@ -101,6 +102,17 @@ class DatavizEditionDialog(BaseEditionDialog, CLASS):
 
         self.setup_ui()
         self.check_form_graph_type()
+        self.check_layer_wfs()
+
+    def check_layer_wfs(self):
+        """ When the layer has changed in the combobox, check if the layer is published as WFS. """
+        layer = self.layer.currentLayer()
+        if not layer:
+            self.show_error(tr('A layer is mandatory.'))
+            return
+
+        not_in_wfs = self.is_layer_in_wfs(layer)
+        self.show_error(not_in_wfs)
 
     def load_collection(self, value) -> None:
         """Load a collection into the table from JSON string."""