+++ lizmap/forms/tooltip_edition.py
@@ -33,6 +33,7 @@ class ToolTipEditionDialog(BaseEditionDialog, CLASS):
         self.config.add_layer_label('colorGeom', self.label_color)
 
         self.layer.setFilters(QgsMapLayerProxyModel.VectorLayer)
+        self.layer.layerChanged.connect(self.check_layer_wfs)
         self.layer.layerChanged.connect(self.fields.set_layer)
         self.fields.set_layer(self.layer.currentLayer())
 
@@ -40,6 +41,17 @@ class ToolTipEditionDialog(BaseEditionDialog, CLASS):
         self.enable_color()
 
         self.setup_ui()
+        self.check_layer_wfs()
+
+    def check_layer_wfs(self):
+        """ When the layer has changed in the combobox, check if the layer is published as WFS. """
+        layer = self.layer.currentLayer()
+        if not layer:
+            self.show_error(tr('A layer is mandatory.'))
+            return
+
+        not_in_wfs = self.is_layer_in_wfs(layer)
+        self.show_error(not_in_wfs)
 
     def enable_color(self):
         if self.display_geometry.isChecked():